requirements.txt
flask~=2.2.5
requests~=2.32.3
werkzeug~=3.0.6
python-dotenv~=1.0.0

numpy~=2.1.3
openai~=1.76.1
# ↓ Add this line
deepface~=0.0.67
pillow~=11.1.0
replicate~=0.22.0
Flask-SQLAlchemy~=3.1.1
Flask-Login~=0.6.3
Flask-WTF~=1.2.1
email-validator~=2.1.1
bcrypt~=4.1.3
authlib~=1.5.2
opencv-python~=4.11.0.86